intra: disable port-fwd; reuse port across uids fails
on Android, firestack unfortunately cannot bind to ports bound by other apps,
which have different uids each. There also doesn't seem to be a namespace-like
isolation even when Block Connections Without VPN (Lockdown) is turned ON;
that is, apps are able to bind to all interfaces (and not just the VPN's)
even in Lockdown.